Remove count of frequency of the first column
Analysis of the data revelead that the numbers in the first column appear, on average, only once. Therefore, the benefit of storing the frequency of the numbers in the first column did not outweight the cost of the increase in memory. Consequently, the solution to the second puzzle now uses a simpler HashMap without the frequency counts of the first column